cert_engine/requirements.txt
# Web Framework & Server
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic-settings==2.2.1
python-multipart==0.0.9

# Database Layer (Async)
sqlalchemy==2.0.29
asyncpg==0.29.0
alembic==1.13.1

# Core Engine: Word Templating
docxtpl==0.16.7


num2words